+++ packages/semver/src/builders/version/utils/workspace.spec.ts
@@ -1,13 +1,11 @@
-import { getPackageFiles, getProjectsRoot } from './workspace';
+import * as fs from 'fs';
+import { callbackify } from 'util';
+
+import { getPackageFiles, getProjectsRoot, readPackageJson } from './workspace';
 
 describe('getPackageFiles', () => {
   it('should resolve projects package.json', () => {
-    expect(
-      getPackageFiles([
-        '/root/packages/a',
-        '/root/packages/b',
-      ])
-    ).toEqual([
+    expect(getPackageFiles(['/root/packages/a', '/root/packages/b'])).toEqual([
       '/root/packages/a/package.json',
       '/root/packages/b/package.json',
     ]);
@@ -30,5 +28,19 @@ describe('getProjectsRoot', () => {
       '/root/packages/cdk/helpers',
       '/root/packages/cdk/operators',
     ]);
+  });
+});
+
+describe('readPackageJson', () => {
+  it('should read package.json', async () => {
+    jest.spyOn(fs, 'readFile').mockImplementation(
+      /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
+      callbackify(jest.fn().mockResolvedValue(`{"version":"2.1.0"}`)) as any
+    );
+
+    const content = await readPackageJson('/root').toPromise();
+    expect(content).toEqual({
+      version: '2.1.0',
+    });
   });
 });
