+++ packages/semver/src/builders/version/utils/workspace.ts
@@ -1,6 +1,8 @@
 import { resolve } from 'path';
+import { Observable } from 'rxjs';
 
 import { SemverConfig } from '../schema';
+import { readJsonFile } from './filesystem';
 
 /**
  * @internal
@@ -22,4 +24,29 @@ export function getProjectsRoot({
   return config.packages.map((projectRoot) =>
     resolve(workspaceRoot, projectRoot)
   );
+}
+
+/**
+ * @internal
+ */
+export function readPackageJson(
+  projectRoot: string
+): Observable<{
+  version?: string;
+}> {
+  return readJsonFile(getPackageJsonPath(projectRoot));
+}
+
+/**
+ * @internal
+ */
+export function getPackageJsonPath(projectRoot: string) {
+  return resolve(projectRoot, 'package.json');
+}
+
+/**
+ * @internal
+ */
+ export function getChangelogPath(projectRoot: string) {
+  return resolve(projectRoot, 'CHANGELOG.md');
 }
